This now passes on windows. The main issue was that __del__ was being called when there was no open script or mesh file??? Could not fathom why, but added an extra test in __del__ to make sure the files were actually open before closing & unlinking.  It may be better to deal with the opening, writing, closing, os.system call, and final unlinking of the script file in getMeshHandler(). That way all manipulations are local, and only the mesh file need be handled in a remote piece of code like __del__.
Not tested on Altix or linux yet.... get to that tonight.

+++ pycad/py_src/gmsh.py
@@ -44,6 +44,9 @@ class Design(design.Design):
     DELAUNAY="iso"
     NETGEN="netgen"
     TETGEN="tetgen"
+
+    __script_file = None
+
     def __init__(self,dim=3,element_size=1.,order=1,keep_files=False):
        """
        initializes the gmsh design
@@ -98,9 +101,11 @@ class Design(design.Design):
         """
         clean up
         """
-        if not self.keepFiles():
-               os.unlink(self.getScriptFileName())
-               os.unlink(self.getMeshFileName())
+        if not self.keepFiles() and (self.scriptFile() != None) :
+            self.scriptFile().close()
+            os.unlink(self.getScriptFileName())
+            os.unlink(self.getMeshFileName())
+
     def getCommandString(self):
         """
         returns the gmsh comand
@@ -128,13 +133,24 @@ class Design(design.Design):
         returns a handle to a mesh meshing the design. In the current implementation 
         a mesh file name in gmsh format is returned.
         """
-        open(self.getScriptFileName(),"w").write(self.getScriptString())
+        f = open(self.getScriptFileName(),"w")
+        setScriptFile(f)
+        f.write(self.getScriptString())
         cmd = self.getCommandString()
         ret = os.system(cmd) / 256
 	if ret > 0:
 	  raise RuntimeError, "Could not build mesh: %s"%cmd
 	else:
           return self.getMeshFileName()
+
+    def setScriptFile(self,f=None):
+        self.__script_file=f
+        return
+
+    def scriptFile(self):
+        return self.__script_file
+
+
 
     def getScriptString(self):
         """
